requirements.txt
openai
Pillow
requests 
configparser
PyQT6